Update list of NRC additions and modifications
Add commands to generate the list of added files and the list of
modified files, for the record. Remove the distinction between major and
minor changes, a list of modified files proves sufficient. Each modified
file can be inspected as suggested here to gauge the substance of the
NRC modification.